requirements.txt
django==4.2.3
requests==2.31.0
djangorestframework==3.14.0
django_bootstrap_icons==0.8.3
selenium==4.11.2
celery==5.1.2
python-dotenv==0.21.1
asgiref==3.6.0
redis==4.5.1
django-extensions==3.2.3
psycopg2-binary==2.9.6
psycopg2==2.9.5
sqlparse==0.4.2
webdriver-manager==4.0.0
undetected-chromedriver==3.5.2
seleniumbase==4.17.8
chromedriver-binary==116.0.5845.96
django-bootstrap-v5==1.0.11
pandas==2.1.0
seaborn==0.12.2
matplotlib==3.8.0
plotly==5.16.1
openai==0.28.1
tiktoken==0.5.1
llama-index==0.8.45.post1